Tests/LibWeb/Text/input/Worker/Worker-performance.html

<script src="../include.js"></script>
<script>
    asyncTest(done => {
        let work = new Worker("worker-performance.js");
        work.onmessage = (evt) => {
            println(`performance global object is available: ${JSON.stringify(evt.data)}`);
            done();
        };
    });
</script>
