Tests/LibWeb/Text/expected/Worker/Worker-performance.txt

performance global object is available: true
